requirements.txt
torch
torchvision
transformers
nltk
sentence-transformers
tqdm
numpy
pillow
fiftyone